refactor: update user UseCase test folder structure
